+++ cmd/zetatool/main.go
@@ -24,6 +24,7 @@ func init() {
 	rootCmd.AddCommand(cli.NewTSSBalancesCMD())
 	rootCmd.AddCommand(cli.NewListChainsCMD())
 	rootCmd.AddCommand(cli.NewListObserversCMD())
+	rootCmd.AddCommand(cli.NewTrackTSSMigrationCMD())
 	rootCmd.PersistentFlags().String(config.FlagConfig, "", "custom config file: --config filename.json")
 	rootCmd.PersistentFlags().
 		Bool(config.FlagDebug, false, "enable debug mode, to show more details on why the command might be failing")